+++ examples/cloud-connect/plugins/bt-business-broadband/config.sh
@@ -1,27 +1,23 @@
 #!/bin/sh
 # shellcheck disable=SC2034
 
-# [UNTESTED!]
-
-# URL/form info based on @SpikeTheLobster's
-# comment over on this Gist:
-# - https://gist.github.com/sscarduzio/05ed0b41d6234530d724#gistcomment-3336485
+# [ ! UNTESTED ! ]
 
 # Network
-BT_SSID="BT-Wifi"
+BT_SSID="BTWi-fi"
 BT_IFACE="apcli0"
 BT_COOKIES="/tmp/bt_cookies.txt"
 BT_PREVIOUS_ATTEMPT="/tmp/bt_previous_attempt.html"
 BT_REBOOT_LOG="${HOME}/bt_reboot_log.txt"
 
 # Portal status + login URLs
-BT_STATUS_URL="https://www.btopenzone.com:8443/home"
-BT_LOGIN_URL="https://www.btwifi.com:8443/ante"
+BT_STATUS_URL="https://www.btwifi.com:8443/home"
+BT_LOGIN_URL="https://www.btwifi.com:8443/ante?partnerNetwork=btb"
 BT_LOGIN_FORM="\
   xhtmlLogon=https://www.btwifi.com:8443/ante&\
-  USERNAME=${__USERNAME__}&\
-  PASSWORD=${__PASSWORD__}&\
-  provider=btoz"
+  username=${__USERNAME__}&\
+  password=${__PASSWORD__}&\
+  provider=business"
 
 # Remove indentation from the form above
 BT_LOGIN_FORM=$(echo "${BT_LOGIN_FORM}"|sed -e 's/  //g')