+++ examples/cloud-connect/plugins/bt-wifi/api.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+# shellcheck disable=SC1090
+
+. "${PLUGIN_PATH}/.secrets"
+. "${PLUGIN_PATH}/config.sh"
+. "${PLUGIN_PATH}/../helpers.sh"
+. "${PLUGIN_PATH}/../bt-api.sh"